Rearrange the test classes of DFS Source.
